Split typedb tests into multiple files
