+++ src/database_builder_libs/stores/typedb/_types.py
@@ -1,5 +1,5 @@
 from typing import Iterator, Mapping, TypedDict
-from typedb.driver import QueryAnswer
+from typedb.driver import ConceptDocumentIterator, ConceptRowIterator, QueryAnswer
 
 
 class RelationRef(TypedDict):
@@ -54,9 +54,9 @@ class EagerQueryAnswer:
         self._is_docs = answer.is_concept_documents()
         self._is_rows = answer.is_concept_rows()
         if self._is_docs:
-            self._docs = answer.as_concept_documents()
+            self._docs = list(answer.as_concept_documents())
         elif self._is_rows:
-            self._rows = answer.as_concept_rows()
+            self._rows = list(answer.as_concept_rows())
         else:
             self._answer = answer
 
@@ -66,13 +66,13 @@ class EagerQueryAnswer:
             raise TypeError("Query did not return concept documents")
         # This is put from an ConceptDocumentIterator to a list and then into a generic Iterator in order to move the data out of the transaction
         # Returned as an iterator for consistency with how the raw Query Answer is used
-        return iter(list(self._docs))
+        return iter(self._docs)
 
     def as_concept_rows(self) -> Iterator:
         """Return the pre-evaluated rows iterator."""
         if not self._is_rows:
             raise TypeError("Query did not return concept rows")
-        return iter(list(self._rows))
+        return iter(self._rows)
 
     def as_raw(self) -> QueryAnswer:
         """Return the underlying raw typeDB.QueryAnswer object, if not evaluated yet."""
@@ -85,7 +85,7 @@ class EagerQueryAnswer:
     def __iter__(self) -> Iterator:
         """Yield items from the eagerly evaluated result."""
         if self._is_docs:
-            return self._docs
+            return iter(self._docs)
         if self._is_rows:
-            return self._rows
+            return iter(self._rows)
         return iter([])
