--- tests/database_builder_libs/stores/test_typedb_store.py
@@ -1,640 +1,0 @@
-import os
-import pytest
-import time
-import socket
-import uuid
-from testcontainers.core.container import DockerContainer
-from typedb.driver import ConceptRowIterator, Credentials, DriverOptions, QueryAnswer, Transaction, TypeDB, TransactionType
-from database_builder_libs.models.node import Node
-from database_builder_libs.stores.typedb.typedb_store import TypeDbDatastore
-
-
-def wait_for_port(host: str, port: int, timeout: float = 60.0):
-    """
-    Continuously tries to connect to the host:port until successful.
-    """
-    start_time = time.time()
-    while True:
-        try:
-            with socket.create_connection((host, port), timeout=1):
-                return
-        except (OSError, ConnectionRefusedError) as exc:
-            if time.time() - start_time > timeout:
-                raise TimeoutError(
-                    f"Port {port} on {host} did not open within {timeout} seconds."
-                ) from exc
-            time.sleep(0.5)
-
-
-@pytest.fixture(scope="module")
-def typedb_container():
-    """
-    Starts a TypeDB 2.29 container and waits for the driver to be ready.
-    """
-    container = (
-        DockerContainer("typedb/typedb:3.8.0")
-        .with_exposed_ports(1729)
-        .with_env("JAVA_TOOL_OPTIONS", "-Xmx1g")
-    )
-
-    container.start()
-    host = "127.0.0.1"
-    port = container.get_exposed_port(1729)
-    address = f"{host}:{port}"
-    credentials = Credentials("admin", "password")
-    driver_options = DriverOptions(is_tls_enabled=False)
-
-    try:
-        print(f"   [DEBUG] Waiting for TCP port {address}...")
-        wait_for_port(host, port, timeout=60)
-
-        max_retries = 30
-        for _ in range(max_retries):
-            try:
-                with TypeDB.driver(address, credentials, driver_options) as driver:
-                    driver.databases.all()
-                    break
-            except Exception as e:
-                print(f"   [DEBUG] Driver connection failed: {e}. Retrying...")
-                time.sleep(1)
-        else:
-            raise RuntimeError("Driver could not connect to TypeDB.")
-
-    except Exception as e:
-        stdout, stderr = container.get_logs()
-        print("\n=== CONTAINER LOGS ===")
-        print(stdout.decode())
-        print(stderr.decode())
-        raise e
-
-    yield address
-    container.stop()
-
-
-@pytest.fixture
-def store(typedb_container):
-    address = typedb_container
-
-    datastore = TypeDbDatastore()
-    schema_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "schema.tql")
-    db_name = f"test_db_{uuid.uuid4().hex}"
-    datastore.connect(
-        {
-            "uri": address,
-            "database": db_name,
-            "schema_path": schema_path,
-            "username": "admin",
-            "password": "password",
-            "tls": False,
-        }
-    )
-
-    yield datastore
-
-    datastore._entity_attr_cache = {}
-    datastore._all_attr_cache = None
-    datastore._key_attr_cache = {}
-
-    if datastore.typedb_driver and datastore.typedb_driver.is_open():
-        try:
-            datastore.typedb_driver.databases.get(db_name).delete()
-        except Exception as e:
-            print(f"Failed to delete test db {db_name}: {e}")
-
-
-@pytest.fixture
-def friendship_store(store: TypeDbDatastore) -> TypeDbDatastore:
-    store.query_schema("""
-        define
-        relation friendship,
-            relates friend,
-            relates friend_of;
-
-        person plays friendship:friend;
-        person plays friendship:friend_of;
-        """)
-    return store
-
-
-@pytest.fixture
-def populated_friendship_store(friendship_store: TypeDbDatastore) -> TypeDbDatastore:
-    friendship_store.query_write(
-        """
-        insert
-            $a isa person, has name "Alice", has email "alice@test.com";
-            $b isa person, has name "Bob", has email "bob@test.com";
-            $f isa friendship, links (friend: $a, friend_of: $b);
-        """
-    )
-    return friendship_store
-
-
-@pytest.fixture
-def alice_node() -> Node:
-    return Node(
-        id="Alice",
-        entity_type="person",
-        key_attribute="name",
-        payload_data={"name": "Alice", "email": "alice@test.com", "age": 25},
-        relations=(),
-    )
-
-
-@pytest.fixture
-def bob_with_friendship_node() -> Node:
-    return Node(
-        id="Bob",
-        entity_type="person",
-        key_attribute="name",
-        payload_data={"name": "Bob", "email": "bob@test.com", "age": 30},
-        relations=(
-            {
-                "type": "friendship",
-                "roles": {
-                    "friend": {
-                        "entity_type": "person",
-                        "key_attr": "name",
-                        "key": "Alice",
-                    },
-                    "friend_of": {
-                        "entity_type": "person",
-                        "key_attr": "name",
-                        "key": "Bob",
-                    },
-                },
-            },
-        ),
-    )
-
-def test_initialization_creates_database_and_schema(store: TypeDbDatastore):
-    rows: ConceptRowIterator = store.query_read("match entity $t sub person;").as_concept_rows()
-    row = next(rows, None)
-    assert row is not None
-    person_type = row.get("t")
-
-    assert person_type is not None
-    assert person_type.get_label() == "person"
-
-def test_requires_connect(typedb_container):
-    store = TypeDbDatastore()
-
-    with pytest.raises(RuntimeError):
-        store.get_nodes("entity=person")
-
-
-def test_save_and_fetch_workflow(store: TypeDbDatastore):
-    store.query_write('insert $p isa person, has name "Alice", has email "alice@test.com";')
-
-    rows = store.query_read("""
-        match 
-        $p isa person, has name "Alice"; 
-        fetch {
-            "name": $p.name,
-            "email": $p.email
-        };
-    """).as_concept_documents()
-
-    results = list(rows)
-    assert len(results) == 1
-
-    assert results[0]["name"] == "Alice"
-    assert results[0]["email"] == "alice@test.com"
-
-
-def test_get_workflow(store: TypeDbDatastore):
-    store.query_write('insert $p isa person, has name "Bob", has age 30;')
-    rows = store.query_read('match $p isa person, has name "Bob";').as_concept_rows()
-
-    results = list(rows)
-
-    assert len(results) > 0
-
-
-def test_update_workflow(store: TypeDbDatastore):
-    store.query_write('insert $p isa person, has name "Charlie", has age 20;')
-
-    update_query = """
-    match
-    $p isa person, has name "Charlie", has age 20;
-    update
-    $p has age 21;
-    """
-    store.query_write(update_query)
-
-    rows = store.query_read('match $p isa person, has name "Charlie"; fetch { "age": $p.age };').as_concept_documents()
-    results = list(rows)
-
-    assert results[0]["age"] == 21
-
-
-def test_delete_workflow(store: TypeDbDatastore):
-    store.query_write('insert $p isa person, has name "Dave";')
-    assert (
-        len(list(
-            store.query_read('match $p isa person, has name "Dave"; fetch { "name": $p.name };'))
-    ))
-
-    store.query_write('match $p isa person, has name "Dave"; delete $p;')
-
-    rows = store.query_read('match $p isa person, has name "Dave"; fetch { "name": $p.name };')
-    results = list(rows)
-
-    assert len(results) == 0
-
-
-def test_get_nodes_by_keyed_filter(store: TypeDbDatastore):
-    # Arrange
-    store.query_write(
-        'insert $p isa person, has name "Alice", has email "alice@test.com", has age 25;'
-    )
-
-    # Act
-    results = store.get_nodes("entity=person&name=Alice")
-
-    # Assert
-    assert len(results) == 1
-    node = results[0]
-
-    assert node.payload_data["name"] == "Alice"
-    assert node.payload_data["email"] == "alice@test.com"
-    assert node.payload_data["age"] == 25
-
-
-def test_get_nodes_returns_empty_when_no_match(store: TypeDbDatastore):
-    results = store.get_nodes("entity=person&name=Nonexistent")
-    assert results == []
-
-
-def test_get_nodes_multiple_matches(store: TypeDbDatastore):
-    store.query_write('insert $p isa person, has name "Bob1", has age 30;')
-    store.query_write('insert $p isa person, has name "Bob2", has age 30;')
-
-    results = store.get_nodes("entity=person&age=30")
-
-    assert len(results) == 2
-    ages = sorted(n.payload_data["age"] for n in results)
-    assert ages == [30, 30]
-
-def test_remove_single_node_by_keyed_filter(store: TypeDbDatastore):
-    store.query_write(
-        'insert $p isa person, has name "Charlie", has email "charlie@test.com";'
-    )
-
-    assert len(store.get_nodes("entity=person&name=Charlie")) == 1
-
-    deleted_count = store.remove_nodes("entity=person&name=Charlie")
-
-    assert deleted_count == 1
-    assert store.get_nodes("entity=person&name=Charlie") == []
-
-
-def test_remove_nodes_refuses_bulk_delete_by_default(store: TypeDbDatastore):
-    store.query_write('insert $p isa person, has name "Dave";')
-    store.query_write('insert $p isa person, has name "Eve";')
-
-    with pytest.raises(ValueError):
-        store.remove_nodes("entity=person")
-
-
-def test_remove_nodes_bulk_delete_with_explicit_flag(store: TypeDbDatastore):
-    store.query_write('insert $p isa person, has name "Frank";')
-    store.query_write('insert $p isa person, has name "Grace";')
-
-    deleted_count = store.remove_nodes(
-        "entity=person",
-        allow_multiple=True,
-    )
-
-    assert deleted_count >= 2
-    assert store.get_nodes("entity=person") == []
-
-
-def test_remove_nodes_only_affects_target_entities(store: TypeDbDatastore):
-    store.query_write('insert $p isa person, has name "Henry";')
-    store.query_write('insert $p isa person, has name "Ivy";')
-
-    store.query_write(
-        'insert $p isa person, has name "Jack";'
-    )
-
-    store.remove_nodes("entity=person&name=Jack")
-
-    remaining = store.get_nodes("entity=person")
-    names = sorted(n.payload_data["name"] for n in remaining)
-
-    assert names == ["Henry", "Ivy"]
-
-
-def test_store_node_inserts_entity(store: TypeDbDatastore, alice_node: Node):
-    store.store_node(alice_node)
-
-    results = store.query_read("""
-        match $p isa person, has email "alice@test.com"; 
-        fetch { $p.* };
-    """).as_concept_documents()
-
-    rows = list(results)
-
-    assert len(rows) == 1
-    person = rows[0]
-
-    assert person["name"] == "Alice"
-    assert person["email"] == "alice@test.com"
-    assert person["age"] == 25
-
-def test_get_nodes_none_returns_all_nodes(store: TypeDbDatastore):
-    # Arrange
-    store.query_write('insert $p isa person, has name "Alice";')
-    store.query_write('insert $p isa person, has name "Bob";')
-    store.query_write('insert $p isa person, has name "Charlie";')
-
-    # Act
-    results = store.get_nodes(None)
-
-    # Assert
-    assert isinstance(results, list)
-    assert len(results) == 3
-
-    names = sorted(n.payload_data["name"] for n in results)
-    assert names == ["Alice", "Bob", "Charlie"]
-
-
-def test_store_node_inserts_relation(friendship_store: TypeDbDatastore, alice_node: Node, bob_with_friendship_node: Node):
-    friendship_store.store_node(alice_node)
-    friendship_store.store_node(bob_with_friendship_node)
-
-    results = friendship_store.query_read(
-        """
-        match
-            $a isa person, has email "alice@test.com";
-            $b isa person, has email "bob@test.com";
-            $f isa friendship, links (friend: $a, friend_of: $b);
-        """
-    ).as_concept_rows()
-
-    assert len(list(results)) == 1
-
-
-def test_get_nodes_with_relations(populated_friendship_store: TypeDbDatastore):
-    nodes = populated_friendship_store.get_nodes("entity=person&email=alice@test.com&include=relations")
-
-    assert len(nodes) == 1
-    node = nodes[0]
-
-    assert len(node.relations) == 1
-    assert node.relations[0]["type"] == "friendship"
-
-
-def test_key_inference_from_schema(store: TypeDbDatastore):
-    store.query_schema("""
-        define
-        attribute username, value string;
-
-        entity account,
-            owns username @key,
-            owns email;
-        """)
-    store.query_write(
-        'insert $a isa account, has username "u1", has email "u1@test.com";'
-    )
-
-    results = store.get_nodes("entity=account")
-
-    assert len(results) == 1
-    node = results[0]
-
-    assert node.key_attribute == "username"
-    assert node.id == "u1"
-
-def test_schema_evolution_new_attribute(store: TypeDbDatastore):
-    store.query_schema("""
-        define
-        attribute nickname, value string;
-
-        person owns nickname;
-        """)
-
-    store.query_write(
-        """
-        insert
-            $p isa person,
-            has name "Alice",
-            has email "alice@test.com",
-            has nickname "Al";
-        """
-    )
-
-    nodes = store.get_nodes("entity=person&email=alice@test.com")
-
-    assert len(nodes) == 1
-    node = nodes[0]
-
-    assert node.payload_data["nickname"] == "Al"
-
-
-def test_relation_with_attributes(store: TypeDbDatastore):
-    tx = store.query_schema("""
-        define
-        relation friendship,
-            relates friend,
-            relates friend_of,
-            owns since;
-
-        attribute since, value integer;
-
-        person plays friendship:friend;
-        person plays friendship:friend_of;
-        """)
-    
-    store.query_write(
-        """
-        insert
-            $a isa person, has name "Alice", has email "alice@test.com";
-            $b isa person, has name "Bob", has email "bob@test.com";
-            $f isa friendship, links (friend: $a, friend_of: $b), has since 2024;
-        """
-    )
-
-    nodes = store.get_nodes("entity=person&email=alice@test.com&include=relations")
-
-    rel = nodes[0].relations[0]
-
-    assert rel["type"] == "friendship"
-    assert rel["attributes"]["since"] == 2024
-
-def test_relation_single_role_player_is_loaded(store: TypeDbDatastore):
-    store.query_schema("""
-    define
-    relation tagged,
-        relates item;
-
-    person plays tagged:item;
-    """)
-
-    store.query_write("""
-        insert
-            $p isa person, has name "Alice", has email "alice@test.com";
-            (item: $p) isa tagged;
-    """)
-
-    nodes = store.get_nodes("entity=person&email=alice@test.com&include=relations")
-
-    assert len(nodes) == 1
-    node = nodes[0]
-
-    assert len(node.relations) == 1
-    assert node.relations[0]["type"] == "tagged"
-
-def test_relation_attributes_are_loaded(store: TypeDbDatastore):
-    store.query_schema("""
-        define
-        relation collaboration,
-            relates contributor,
-            relates project,
-            owns contribution_role;
-
-        attribute contribution_role, value string;
-
-        person plays collaboration:contributor;
-        person plays collaboration:project;
-        """)
-
-    store.query_write("""
-        insert
-            $a isa person, has name "Alice", has email "alice@test.com";
-            $b isa person, has name "Bob", has email "bob@test.com";
-            (contributor: $a, project: $b) isa collaboration, has contribution_role "author";
-    """)
-
-    nodes = store.get_nodes("entity=person&email=alice@test.com&include=relations")
-
-    rel = nodes[0].relations[0]
-
-    assert rel["attributes"]["contribution_role"] == "author"
-
-def test_relations_not_duplicated(populated_friendship_store: TypeDbDatastore):
-    nodes = populated_friendship_store.get_nodes("entity=person&email=alice@test.com&include=relations")
-
-    assert len(nodes[0].relations) == 1
-
-def test_store_node_relation_idempotent(friendship_store: TypeDbDatastore, alice_node: Node, bob_with_friendship_node: Node):
-    """
-    Store the same node again to verify that the operation is idempotent.
-    Re-inserting an existing node should not create duplicates or duplicate relations. 
-    This ensures that calling `store_node` multiple times with the same data leaves the graph in the same state.
-    """
-    friendship_store.store_node(alice_node)
-    friendship_store.store_node(bob_with_friendship_node)
-
-    # insert again
-    friendship_store.store_node(bob_with_friendship_node)
-
-    nodes = friendship_store.get_nodes("entity=person&name=Bob&include=relations")
-
-    assert len(nodes[0].relations) == 1
-
-
-def test_multiple_relations_retrieval(store: TypeDbDatastore):
-    store.query_schema("""
-        define
-        relation friendship,
-            relates friend,
-            relates friend_of;
-
-        relation tagged,
-            relates item;
-
-        person plays friendship:friend;
-        person plays friendship:friend_of;
-        person plays tagged:item;
-        """)
-
-    store.query_write("""
-        insert
-            $a isa person, has name "Alice", has email "alice@test.com";
-            $b isa person, has name "Bob", has email "bob@test.com";
-            $f isa friendship, links (friend: $a, friend_of: $b);
-            (item: $a) isa tagged;
-    """)
-
-
-    nodes = store.get_nodes("entity=person&email=alice@test.com&include=relations")
-
-    assert len(nodes) == 1
-    node = nodes[0]
-
-    assert len(node.relations) == 2
-    relation_types = sorted(r["type"] for r in node.relations)
-    assert relation_types == ["friendship", "tagged"]
-
-
-def test_get_relations_returns_all_when_none(populated_friendship_store: TypeDbDatastore):
-    relations = populated_friendship_store.get_relations(None)
-    assert len(relations) >= 1
-    types = [r["type"] for r in relations]
-    assert "friendship" in types
-
-
-def test_get_relations_by_type(populated_friendship_store: TypeDbDatastore):
-    relations = populated_friendship_store.get_relations("relation=friendship")
-    assert len(relations) == 1
-    rel = relations[0]
-    assert rel["type"] == "friendship"
-    assert "friend" in rel["roles"]
-    assert "friend_of" in rel["roles"]
-    assert rel["roles"]["friend"]["key"] == "Alice"
-    assert rel["roles"]["friend_of"]["key"] == "Bob"
-
-
-def test_get_relations_with_attributes(store: TypeDbDatastore):
-    store.query_schema("""
-        define
-        relation friendship,
-            relates friend,
-            relates friend_of,
-            owns since;
-
-        attribute since, value integer;
-
-        person plays friendship:friend;
-        person plays friendship:friend_of;
-        """)
-    
-    store.query_write(
-        """
-        insert
-            $a isa person, has name "Alice", has email "alice@test.com";
-            $b isa person, has name "Bob", has email "bob@test.com";
-            $f isa friendship, links (friend: $a, friend_of: $b), has since 2024;
-        """
-    )
-
-    relations = store.get_relations("relation=friendship&since=2024")
-    assert len(relations) == 1
-    rel = relations[0]
-    assert rel["type"] == "friendship"
-    assert rel["attributes"]["since"] == 2024
-
-
-def test_get_relations_returns_empty_when_no_match(store: TypeDbDatastore):
-    store.query_schema("""
-        define
-        relation friendship,
-            relates friend,
-            owns since;
-
-        attribute since, value integer;
-
-        person plays friendship:friend;
-        """)
-        
-    store.query_write(
-        """
-        insert
-            $a isa person, has name "Alice", has email "alice@test.com";
-            $f isa friendship, links (friend: $a), has since 2024;
-        """
-    )
-    
-    relations = store.get_relations("relation=friendship&since=2099")
-    assert relations == []
