+++ tests/database_builder_libs/stores/typedb/schema.tql
